requirements.txt
discord.py==0.16.12
PyYAML==3.12
